+++ resolwe_bio/tests/workflows/test_atac_seq.py
@@ -31,11 +31,42 @@ class AtacSeqWorkflowTestCase(BioProcessTestCase):
                 "reads": reads.id,
                 "genome": bowtie2_index.id,
                 "promoter": promoters.id,
-                "settings": {"bedgraph": False},
             },
         )
         for data in Data.objects.all():
             self.assertStatus(data, Data.STATUS_DONE)
-        macs2 = Data.objects.last()
+
+    @tag_process("workflow-atac-seq-beta")
+    def test_beta_atac_seq_workflow(self):
+        with self.preparation_stage():
+            inputs = {
+                "src": "mm10_chr17_44M-45M.fa.gz",
+                "species": "Mus musculus",
+                "build": "mm10",
+            }
+            ref_seq = self.run_process("upload-fasta-nucl", inputs)
+            bowtie2_index = self.run_process("bowtie2-index", {"ref_seq": ref_seq.id})
+            reads = self.prepare_paired_reads(
+                mate1=["atac_R1.fastq.gz"], mate2=["atac_R2.fastq.gz"]
+            )
+            inputs = {
+                "src": "promoters_mm10_chr17_subregion.bed",
+                "species": "Mus musculus",
+                "build": "mm10",
+            }
+            promoters = self.run_process("upload-bed", inputs)
+
+        self.run_process(
+            "workflow-atac-seq-beta",
+            {
+                "reads": reads.id,
+                "genome": bowtie2_index.id,
+                "promoter": promoters.id,
+                "macs2_options": {"bedgraph": False},
+            },
+        )
+        for data in Data.objects.all():
+            self.assertStatus(data, Data.STATUS_DONE)
+        macs2 = Data.objects.get(process__slug="macs2-callpeak")
         self.assertFile(macs2, "chip_qc", "atac_seq_report.txt")
         self.assertFile(macs2, "case_prepeak_qc", "atac_seq_prepeak_report.txt")